* OLDateTime changed: NullFlag: the string field has been replaced with an FHasValue field, which is either Boolean or String depending on the Delphi version. In newer versions of Delphi, the initialization operator is used, so the Boolean flag is automatically set to False.
Operators Equal, NotEqual, GreaterThan, GreaterThanOrEqual, LessThan, and LessThanOrEqual now return Boolean instead of OLBoolean. This enables short-circuit evaluation in complex logical operations.

For example:
if (i = 5) and (MyFunc() < OLDate.Today()) then...

Previously, MyFunc() was always executed even when i was not equal to 5. This was because (MyFunc() < 50) returned OLBoolean, which prevented short-circuiting.